[BE] refactor: IntegrationTest 하나로 통합 #205
