+++ backend/src/test/java/com/onair/hearit/fixture/IntegrationTest.java
@@ -1,26 +1,39 @@
 package com.onair.hearit.fixture;
 
+import static org.springframework.restdocs.restassured.RestAssuredRestDocumentation.documentationConfiguration;
+
 import io.restassured.RestAssured;
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.specification.RequestSpecification;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.restdocs.RestDocumentationContextProvider;
+import org.springframework.restdocs.RestDocumentationExtension;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
 
 @Sql("/dbclean.sql")
 @ActiveProfiles("integration-test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ExtendWith(RestDocumentationExtension.class)
 public abstract class IntegrationTest {
 
     @LocalServerPort
     int port;
 
     @Autowired
-    DbHelper dbHelper;
+    public DbHelper dbHelper;
+
+    protected RequestSpecification spec;
 
     @BeforeEach
-    void setupPort() {
+    void setUp(RestDocumentationContextProvider provider) {
         RestAssured.port = port;
+        this.spec = new RequestSpecBuilder()
+                .addFilter(documentationConfiguration(provider))
+                .build();
     }
 }
